+++ opencontainers/image/v1/__init__.py
@@ -21,8 +21,8 @@ from .annotations import (
 )
 
 from .config import (
-    ImageConfig
-    RootFS
+    ImageConfig,
+    RootFS,
     Image
 )