+++ opencontainers/struct.py
@@ -6,6 +6,7 @@
 # with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from opencontainers.logger import bot
+import copy
 import json
 
 
@@ -32,6 +33,21 @@ class StructAttr(object):
         self.required = required
         self.jsonName = jsonName or name
         self.omitempty = omitempty
+
+    def __str__(self):
+        return "<opencontainers.struct.StructAttr-%s:%s>" %(self.name, self.value)
+
+    def __repr__(self):
+        return self.__str__()
+
+    def set(self, value):
+        '''set a new value, and validate the type. Return true if set
+        '''
+        if self.validate_type(value):
+            self.value = value
+            return True
+        return False
+
 
     def validate_type(self, value):
         '''ensure that an attribute is of the correct type. If we are given
@@ -67,7 +83,6 @@ class Struct(object):
     '''
     attrs = {}
 
-
     def newAttr(self, name, attType, required=False, jsonName=None, omitempty=True):
         '''add a new attribute, including a name, json key to dump,
            type, and if required. We don't need a value here
@@ -82,7 +97,18 @@ class Struct(object):
         '''
         if name in self.attrs:
             bot.exit("%s has already been added." % name)
-        self.attrs[name] = StructAttr(name, attType, required, jsonName, omitempty)
+        self.attrs[name] = StructAttr(name=name, 
+                                      attType=attType,
+                                      required=required,
+                                      jsonName=jsonName, 
+                                      omitempty=omitempty)
+
+    def _clear_values(self):
+        '''if a load is done, we remove previously loaded values for any
+           attributes
+        '''
+        for name, att in self.attrs.items():
+            self.attrs[name].value = None        
 
 
     def to_dict(self):
@@ -103,7 +129,7 @@ class Struct(object):
                     result[att.jsonName] = att.value
             return result
 
-    def to_json(self)L
+    def to_json(self):
         '''get the dictionary of a struct and return pretty printed json
         '''
         result = self.to_dict()
@@ -120,16 +146,46 @@ class Struct(object):
 
         attr = self.attrs[name]
 
-        if not value:
-            bot.warning("adding empty value to %s." % name)
-
-        if not attr.validate_type(value)
-            bot.exit("%s must be type %s." %(name, att.attType))
-
-        self.attrs[name].value = value
+        # Don't validate the type if provided is empty
+        if value:
+            if not attr.set(value):
+                bot.exit("%s must be type %s." %(name, attr.attType))
 
 
-    # TODO a load function
+    def load(self, content):
+        '''given a dictionary load into its respective object
+        '''
+        if not isinstance(content, dict):
+            bot.exit("Please provide a dictionary to load.")
+
+        # Validate the content, only clear if it's valid
+        attrs = copy.deepcopy(self.attrs)
+
+        # Look up attributes based on jsonKey
+        lookup = self.generate_json_lookup()
+
+        # Go through each value for the content, validate
+        for key, value in content.items():
+            att = lookup.get(key)
+            if not att:
+                bot.exit("%s is not a valid json attribute." % key)
+
+        # If we get here, all parameters are valid, replace
+        self._clear_values()
+        for key, value in content.items():
+            att = lookup.get(key)
+            att.set(value)
+
+
+    def generate_json_lookup(self):
+        '''based on the attributes, generate a jsonName lookup object.
+           keys are jsonNames we find in the wild, names are attribute names.
+        '''
+        lookup = dict()
+        for name, att in self.attrs.items():
+            lookup[att.jsonName] = att
+        return lookup
+
 
     def validate(self):
         '''validate goes through each attribute, and ensure that it is of the
@@ -145,7 +201,7 @@ class Struct(object):
 
             # The attribute must match its type
             if not isinstance(att.value, att.attType):
-                bot.error("%s should be type %s" %(name, att.Type))
+                bot.error("%s should be type %s" %(name, att.attType))
                 return False
 
-       return True
+        return True